requirements.txt
python-telegram-bot==20.8
wakeonlan==3.1.0